requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
certifi==2020.6.20
Django==3.0.3
django-bootstrap4==2.2.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.1
django-filter==2.3.0
django-google-maps==0.12.1
django-js-asset==1.2.2
Pillow==7.2.0
psycopg2==2.8.5
python-decouple==3.3
pytz==2020.1
sorl-thumbnail==12.6.3
soupsieve==2.0.1
sqlparse==0.3.1
wincertstore==0.2
